python3 compat (float division)

+++ gui/tools/read_burcat.py
@@ -37,7 +37,7 @@ def parse_section(section):
             pass
     else:
         comment = '\n'.join(comment_block)
-        for i in range(len(data_block)/4):
+        for i in range(int(len(data_block)/4)):
             l1, l2, l3, l4 = data_block[i*4 : i*4+4]
             name = l1[:18].strip()
             phase = l1[44]
